Move entities to separate folder
