+++ domain/entity/ActionSpace.ts
