+++ domain/entity/Player.ts
