+++ domain/gatherWood.unit.ts
@@ -1,9 +1,9 @@
 import { expect } from "chai";
 import { GatherWood } from "./gatherWood";
-import { Dwarf, Player } from "./Player";
-import { ActionSpace, ActionSpaceId } from "./ActionSpace";
-import { Resources, ResourceType } from "./Resources";
-import { Game } from "./Game";
+import { Dwarf, Player } from "./entity/Player";
+import { ActionSpace, ActionSpaceId } from "./entity/ActionSpace";
+import { Resources, ResourceType } from "./entity/Resources";
+import { Game } from "./entity/Game";
 import { buildBaseObjects, expectMutationsOfType, shouldPlaceDwarf } from "./testUtils";
 
 describe("Gather wood", () => {
